backend/requirements.txt
fastapi[all]
uvicorn
jinja2
weasyprint
pandas
psycopg2-binary
pyyaml